Repo details loaded.

+++ GithubRepo/Models/RepositoryDataModel.swift
@@ -11,7 +11,7 @@ import ObjectMapper
 
 class RepositoryDataModel: NSObject, Mappable {
     
-    var total_count: Int64?
+    var total_count: Int?
     var incomplete_results: Bool?
     var items: [RepositoryModel]?
